Refactor code for improved clarity and functionality
- Removed unused linting command from composer.json.
- Updated rector.php to include new Laravel set and skip rules for better code quality.
- Modified CreateNewUser action to use query builder for user creation.
- Enhanced PasswordValidationRules to improve type hinting.
- Updated various console commands to utilize the #[Override] attribute for better clarity.
- Refactored DashboardController and GameController to extend the base Controller class directly.
- Improved Game model by adding new query scopes and updating relationships for better performance.
- Cleaned up GameRequest and GameRequestVote models for consistency in type hinting.
- Enhanced error handling in services and data providers for better robustness.
- Updated RssFeedFetcher to improve publication date parsing logic.

+++ database/factories/GameFactory.php
@@ -28,8 +28,8 @@ final class GameFactory extends Factory
             'cover_image' => fake()->optional(0.5)->imageUrl(400, 600, 'games'),
             'developer' => fake()->optional(0.8)->company(),
             'publisher' => fake()->optional(0.6)->company(),
-            'genres' => fake()->randomElements(['RPG', 'Action', 'Adventure', 'Strategy', 'Racing', 'Shooter', 'Puzzle'], (int) fake()->numberBetween(1, 3)),
-            'platforms' => fake()->randomElements(['PC', 'PlayStation', 'Xbox', 'Switch'], (int) fake()->numberBetween(1, 3)),
+            'genres' => fake()->randomElements(['RPG', 'Action', 'Adventure', 'Strategy', 'Racing', 'Shooter', 'Puzzle'], fake()->numberBetween(1, 3)),
+            'platforms' => fake()->randomElements(['PC', 'PlayStation', 'Xbox', 'Switch'], fake()->numberBetween(1, 3)),
             'release_date' => fake()->optional(0.8)->dateTimeBetween('-1 year', '+1 year'),
             'release_status' => fake()->randomElement(ReleaseStatus::cases()),
         ];
